Moved CheatManager in dedicated folder
